making new poll ui
