+++ client/src/types/clerk.d.ts
@@ -1,0 +1,10 @@
+import { UserRole } from "@/lib/types";
+
+declare module "@clerk/clerk-react" {
+  interface UserResource {
+    publicMetadata: {
+      role?: UserRole;
+      planType?: string;
+    };
+  }
+}
